+++ lib/mangaloom_parser.dart
@@ -10,13 +10,9 @@ export 'src/models/genre.dart';
 // Parsers
 export 'src/parsers/parser_base.dart';
 export 'src/parsers/id/shinigami_parser.dart';
-export 'src/parsers/id/comicsans_parser.dart';
 export 'src/parsers/id/mangaplus_parser.dart';
 export 'src/parsers/en/mangapark_parser.dart';
 export 'src/parsers/id/webtoon_parser.dart';
-export 'src/parsers/en/batoto_parser.dart';
-export 'src/parsers/id/komiklu_parser.dart';
 export 'src/parsers/id/komiku_parser.dart';
 export 'src/parsers/lib/natsu_parser.dart';
-export 'src/parsers/id/kiryuu_parser.dart';
 export 'src/parsers/id/ikiru_parser.dart';